refactor: date, time options 생성하는 로직 useSetReminder hook 에서 분리
